+++ src/components/patterns/statsCarousel/statsCarouselSlide.stories.ts
@@ -1,10 +1,10 @@
 import { html } from 'lit';
-import { createOptionsArray } from '../../common/helpers/helpers';
+import { createOptionsArray } from '@kyndryl-design-system/shidoka-foundation/common/helpers/storybook';
 import './index';
 import { STAT_CAROUSEL_SLIDE_SYMBOL_POSITIONS, STAT_CAROUSEL_SLIDE_SYMBOL_ALIGNMENT } from './defs';
 
 export default {
-	title: 'Components/Stats Carousel/Stats Carousel Slide',
+	title: 'Patterns/Stats Carousel/Stats Carousel Slide',
 	component: 'kd-stats-carousel-slide',
 	argTypes: {
 		stat: {
@@ -25,9 +25,9 @@ export default {
 };
 
 const args = {
-	stat: '100',
-	symbol: '$',
-	symbolPlacement: STAT_CAROUSEL_SLIDE_SYMBOL_POSITIONS.BEFORE,
+	stat: '$100',
+	symbol: 'M',
+	symbolPlacement: STAT_CAROUSEL_SLIDE_SYMBOL_POSITIONS.AFTER,
 	symbolAlignment: STAT_CAROUSEL_SLIDE_SYMBOL_ALIGNMENT.SUPERSCRIPT,
 }